Moving Cpb from deprecated schema to Code
